// src/api/stetsom/model/pagePayload.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { I18nString } from "./i18nString";
import type { PageBlock } from "./pageBlock";

export interface PagePayload {
  id: string;
  slug: string;
  title: I18nString;
  blocks: PageBlock[];
  updated_at: string;
}